+++ src/bxai/parametric/__init__.py
@@ -1,7 +1,9 @@
 from bxai.parametric.bart_importance import BARTImportance
+from bxai.parametric.correlation import BayesianCorrelation
 from bxai.parametric.shrinkage_pip import ShrinkagePIP
 
 __all__ = [
     "ShrinkagePIP",
     "BARTImportance",
+    "BayesianCorrelation",
 ]